fix: extract advisor shared types to client-safe module
Importing PENDING_ACTION_PART_TYPE from advisor-tools.ts in a 'use client'
component pulled server-only modules into the client bundle, causing a
circular dependency TDZ error at runtime.

Move the constant and PendingAction types to advisor-types.ts (no server
imports). Client component now imports from there; advisor-tools.ts re-exports
for server-side callers.
